Split ScopedLiftActor out into its own file.
